introduce a new namespace Smdn.Devices.Mcp2221A.Peripherals.Gpio to separate nested types within the Mcp2221A class
