--- src/Smdn.Devices.Mcp2221A/Smdn.Devices.Mcp2221A/Mcp2221A.GPs.InterruptDetection.cs
@@ -1,22 +1,0 @@
-// SPDX-FileCopyrightText: 2021 smdn <smdn@smdn.jp>
-// SPDX-License-Identifier: MIT
-
-using System.Threading;
-using System.Threading.Tasks;
-
-namespace Smdn.Devices.Mcp2221A;
-
-#pragma warning disable IDE0040
-partial class Mcp2221A {
-#pragma warning restore IDE0040
-  internal interface IInterruptDetectionFunctionality {
-    ValueTask ConfigureAsInterruptDetectionAsync(
-      CancellationToken cancellationToken = default
-    );
-    void ConfigureAsInterruptDetection(
-      CancellationToken cancellationToken = default
-    );
-#if __FUTURE_VERSION
-#endif
-  }
-}
